refactor: built the project structure in accordance with FSD, feat: set up absolute imports
