+++ src/app/providers/ThemeProvider/lib/useTheme.ts
@@ -1,5 +1,5 @@
-import {LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext} from "./ThemeContext";
 import {useContext} from "react";
+import {LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext} from "app/providers/ThemeProvider/ui/ThemeContext";
 
 interface UseThemeResult {
     toggleTheme: () => void;
@@ -8,7 +8,7 @@ interface UseThemeResult {
 
 export const useTheme = ():UseThemeResult => {
     const {theme, setTheme} = useContext(ThemeContext);
-
+debugger
     const toggleTheme = () => {
         const newTheme = theme === Theme.DARK ? Theme.LIGHT : Theme.DARK
         setTheme(newTheme);
