+++ src/app/styles/themes/dark.scss
@@ -1,0 +1,6 @@
+.app.dark {
+    --bg-color: #090949;
+
+    --primary-color: #049604;
+    --secondary-color: #04ff04;
+}
