+++ src/app/styles/themes/normal.scss
@@ -1,0 +1,6 @@
+:root {
+    --bg-color: #e8e8ea;
+
+    --primary-color: #0232c2;
+    --secondary-color: #0449e0;
+}
