+++ src/app/styles/variables/global.scss
@@ -1,5 +1,5 @@
 :root {
-  --font-family-main: Consolas, 'Times New Roman', Serif;
+  --font-family-main: Consolas, "Times New Roman", Serif;
 
   --font-size-m: 16px;
   --font-line-m: 24px;
