--- src/components/Counter.module.scss
@@ -1,4 +1,0 @@
-.btn {
-    padding: 20px;
-    color: green;
-}
