--- src/index.scss
@@ -1,13 +1,0 @@
-@import "styles/reset";
-@import "styles/variables/global";
-@import "styles/themes/normal";
-@import "styles/themes/dark";
-
-.app {
-  font: var(--font-m);
-
-  background: var(--bg-color);
-  color: var(--primary-color);
-
-  min-height: 100vh;
-}
