+++ src/pages/AboutPage/index.tsx
@@ -1,0 +1,5 @@
+import {AboutPageAsync} from "./ui/AboutPage.async";
+
+export {
+    AboutPageAsync as AboutPage
+}
