+++ src/pages/MainPage/index.tsx
@@ -1,0 +1,5 @@
+import {MainPageAsync} from "./ui/MainPage.async";
+
+export {
+    MainPageAsync as MainPage
+}
