--- src/styles/themes/dark.scss
@@ -1,6 +1,0 @@
-.app.dark {
-  --bg-color: #090949;
-
-  --primary-color: #049604;
-  --secondary-color: #04ff04;
-}
