--- src/styles/themes/normal.scss
@@ -1,6 +1,0 @@
-.app.light {
-  --bg-color: #e8e8ea;
-
-  --primary-color: #041c65;
-  --secondary-color: #0449e0;
-}
