+++ webpack.config.ts
@@ -5,10 +5,12 @@ import {BuildEnv, BuildPaths} from "./config/build/types/config";
 
 
 export default (env: BuildEnv) =>{
+
     const paths: BuildPaths = {
         entry: path.resolve(__dirname, 'src', 'index.tsx'),
         build: path.resolve(__dirname, 'build'),
-        html: path.resolve(__dirname, 'public', 'index.html')
+        html: path.resolve(__dirname, 'public', 'index.html'),
+        src: path.resolve(__dirname, 'src',)
     }
     const mode = env.mode || 'development'
     const PORT =  env.port || 3000
